deleted old RecordViewModel and moved new one from outside
