+++ src/main/java/framework/buttons/ButtonSearch.java
@@ -1,7 +1,6 @@
 package framework.buttons;
 
 import interface_adaptors.SearchResultsViewModel;
-import interface_adaptors.playlist_ia.RecordViewModel;
 
 
 import javax.imageio.ImageIO;
