+++ src/main/java/framework/items/FriendsCollectionItem.java
@@ -2,7 +2,6 @@ package framework.items;
 
 import entities.user_entities.User;
 import interface_adaptors.SearchResultsViewModel;
-import interface_adaptors.playlist_ia.RecordViewModel;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
